tp2/requirements.txt
python-dotenv
groq
pinecone
langchain-text-splitters
langchain-pinecone
langchain-openai